Former-commit-id: 7b1b43ed469e4e3f5849ebab3c3fef2b174afd37 [formerly 65c2afa06454c3f17e57369130a579dbb616115e] [formerly caf7f8db6f6b533872fab793a86ec3b7aa7fcbcd [formerly 5ca6b0cee9f192d61bdd95a7fb47881ca163ab09]]
Former-commit-id: e822efdcacceda8d3cb815304253ed774c4377ac [formerly cbd98438a6e6de876bf8810d59f335cadc4f3a3a]
Former-commit-id: f47f20ff016a62bdbd4db835a2701745574bd8ac

+++ src/util/io.cc
@@ -197,7 +197,7 @@ namespace FT{
          */
         void load_partial_longitudinal(const std::string & path,
                                std::map<string, std::pair<vector<ArrayXd>, vector<ArrayXd> > > &Z,
-                               char sep, const vector<long>& idx)
+                               char sep, const vector<int>& idx)
         {
             /* loads data from the longitudinal file, with idx providing the id numbers of each row in
              * the main data (X and y).
@@ -207,7 +207,7 @@ namespace FT{
             /* cout << idx.size() << " indices\n"; */
             /* for (unsigned i = 0; i<idx.size(); ++i) */
             /*     cout << i << "," << idx[i] << "\n"; */
-            std::unordered_set<long> idSet; //(idx.begin(), idx.end());
+            std::unordered_set<int> idSet; //(idx.begin(), idx.end());
             /* idSet.insert(idx.begin(), idx.end()); */
             // write out info about unordered_set
             
@@ -313,7 +313,7 @@ namespace FT{
                 if(!firstKey.compare(""))
                     firstKey = name;
                 
-                long sID = std::stol(sampleNo);
+                int sID = std::stol(sampleNo);
                 if(idSet.find(sID) != idSet.end())  // if the sample ID is to be included, store it
                 /* if(in(idx,sNo))  // if the sample ID is to be included, store it */
                 {
